Connect and pull/get data from a DHIS2 instance

To create a DHIS2 connection, you need a DHIS2 base URL, username, password, and an API version The R6 Class called `Dhis2r` representing a DHIS2 instance connection

Format

An R6 class called Dhis2r.

Details

You can use a DHIS2 instance connection to get data several times without needing to manually supply your user credentials on each API call.

Public fields

request_sent

The request used to perform an API call

name

Name of the user

access_rights

The access rights the user has on the DHIS2 instance

account_info

Information of the logged account credentials

Methods

Public methods

Dhis2r$new()

Dhis2r$get_user_info()

Dhis2r$get_metadata()

Dhis2r$get_metadata_fields()

Dhis2r$get_analytics()

Dhis2r$get_any_analytic()

Dhis2r$clone()

Method new()

Create a connection to a DHIS2 instance using basic authentication

Usage

Dhis2r$new(
  base_url,
  username = NULL,
  password = NULL,
  api_token = NULL,
  api_version = NULL,
  api_version_position = c("after", "before")
)

Arguments

base_url

Base url e.g https://play.dhis2.org/

username

Registered username e.g "admin"

password

Registered password e.g "district"

api_token

Personal Access Token (PAT) to use instead of username and password

api_version

The api version e.g "33"

api_version_position

position where the api_version is after or before in web API url i.e /api/

Returns

A new `Dhis2r` object

Method get_user_info()

Get information of the logged in user

Usage

Dhis2r$get_user_info()

Returns

A vector

Method get_metadata()

Get metadata about any available resource from a DHIS2 instance e.g "dataElements", "organisationUnits", "indicators", "periodTypes"

Usage

Dhis2r$get_metadata(endpoint = NULL, fields = c("name", "id"))

Arguments

endpoint

a resource, get the available resources using `get_metadata()` without any arguments

fields

The specific columns to be return in the dataframe e.g c("name","id")

Returns

A data frame

Method get_metadata_fields()

Get all possible fields for a specific metadata resource from a DHIS2 instance

Usage

Dhis2r$get_metadata_fields(endpoint)

Arguments

endpoint

a resource, get the available resources using `get_metadata()` without any arguments

Returns

A vector of all possible fields for a specific metadata

Method get_analytics()

Get all possible analytics resources from a DHIS2 instance i.e

Usage

Dhis2r$get_analytics(
  analytic,
  org_unit,
  period,
  output_scheme = c("UID", "NAME")
)

Arguments

analytic

vector of ID of specific analytic(s) from a DHIS2 instance

org_unit

vector of ID of specific organisation unit(s) from a DHIS2 instance

period

vector of relative or fixed periods from a DHIS2 instance

output_scheme

Output type ID or Names of fields

Returns

A data frame of the analytics resource

Method get_any_analytic()

Get all any analytics resource from a DHIS2 instance to cater for long DHIS2 favorites

Usage

Dhis2r$get_any_analytic(endpoint_url)

Arguments

endpoint_url

string part of Analytic(s) from a DHIS2 instance api endpoint starting from 'analytics.json?dimension='

Returns

A data frame of the analytics resource

Method clone()

The objects of this class are cloneable with this method.

Usage

Dhis2r$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Load dhis2r
library(dhis2r)
# connect to the DHIS2 instance
dhis2_play_connection <- Dhis2r$new(base_url = "https://play.im.dhis2.org/stable-2-40-5",
username = "admin",
password = "district")

# get all the available resources
dhis2_play_connection$get_metadata()
#> [1] "https://play.im.dhis2.org/stable-2-40-5/api/resources"
#> # A tibble: 89 × 4
#>    displayName                     singular                    plural      href 
#>    <chr>                           <chr>                       <chr>       <chr>
#>  1 Program Rules                   programRule                 programRul… http…
#>  2 Aggregate Data Exchanges        aggregateDataExchange       aggregateD… http…
#>  3 Tracked Entity Instance Filters trackedEntityInstanceFilter trackedEnt… http…
#>  4 Program Stages                  programStage                programSta… http…
#>  5 Options                         option                      options     http…
#>  6 Program Sections                programSection              programSec… http…
#>  7 Proposals                       proposal                    proposals   http…
#>  8 Option Group Sets               optionGroupSet              optionGrou… http…
#>  9 Indicator Groups                indicatorGroup              indicatorG… http…
#> 10 Job Configurations              jobConfiguration            jobConfigu… http…
#> # ℹ 79 more rows

# get organisation Units with the default fields i.e c("name","id")

dhis2_play_connection$get_metadata(endpoint = "organisationUnits")
#> [1] "https://play.im.dhis2.org/stable-2-40-5/api/organisationUnits?fields=name%2Cid"
#> # A tibble: 5 × 1
#>   `response_data[[1]]`
#>   <named list>        
#> 1 <int [1]>           
#> 2 <int [1]>           
#> 3 <int [1]>           
#> 4 <chr [1]>           
#> 5 <int [1]>           

# get a vector of all possible fields of a organisation unit resource
dhis2_play_connection$get_metadata_fields(endpoint = "organisationUnits")
#> [1] "https://play.im.dhis2.org/stable-2-40-5/api/organisationUnits?fields=%3Aall&paging=true&pageSize=1"
#>  [1] "access"                 "ancestors"              "attributeValues"       
#>  [4] "children"               "created"                "createdBy"             
#>  [7] "dataSets"               "dimensionItem"          "dimensionItemType"     
#> [10] "displayFormName"        "displayName"            "displayShortName"      
#> [13] "externalAccess"         "favorite"               "favorites"             
#> [16] "href"                   "id"                     "lastUpdated"           
#> [19] "lastUpdatedBy"          "leaf"                   "legendSets"            
#> [22] "level"                  "name"                   "openingDate"           
#> [25] "organisationUnitGroups" "parent"                 "path"                  
#> [28] "programs"               "sharing"                "shortName"             
#> [31] "translations"           "user"                   "userAccesses"          
#> [34] "userGroupAccesses"      "users"                 

# get organisation Units with additional fields i.e c("name","id", "level")

dhis2_play_connection$get_metadata(endpoint = "organisationUnits",
fields =  c("name","id", "level"))
#> [1] "https://play.im.dhis2.org/stable-2-40-5/api/organisationUnits?fields=name%2Cid%2Clevel"
#> # A tibble: 5 × 1
#>   `response_data[[1]]`
#>   <named list>        
#> 1 <int [1]>           
#> 2 <int [1]>           
#> 3 <int [1]>           
#> 4 <chr [1]>           
#> 5 <int [1]>           

dhis2_play_connection$get_analytics(analytic = c("Uvn6LCg7dVU"),
org_unit =   c("O6uvpzGd5pu", "fdc6uOvgoji"),
period = "LAST_12_MONTHS",
 output_scheme = "NAME")
#> [1] "https://play.im.dhis2.org/stable-2-40-5/api/analytics?dimension=dx%3AUvn6LCg7dVU&dimension=ou%3AO6uvpzGd5pu%3Bfdc6uOvgoji&dimension=pe%3ALAST_12_MONTHS&outputIdScheme=NAME"
#> # A tibble: 22 × 4
#>    analytic       org_unit period        value
#>    <fct>          <fct>    <chr>         <dbl>
#>  1 ANC 1 Coverage Bo       November 2023  194.
#>  2 ANC 1 Coverage Bo       December 2023  111.
#>  3 ANC 1 Coverage Bo       January 2024   127.
#>  4 ANC 1 Coverage Bo       February 2024  133.
#>  5 ANC 1 Coverage Bo       March 2024     124.
#>  6 ANC 1 Coverage Bo       April 2024     133.
#>  7 ANC 1 Coverage Bo       May 2024       154.
#>  8 ANC 1 Coverage Bo       June 2024      142.
#>  9 ANC 1 Coverage Bo       July 2024      171.
#> 10 ANC 1 Coverage Bo       August 2024    144.
#> # ℹ 12 more rows